// src/components/proTable/table.types.ts
type ProAsyncEventType = () => {
  resolve: () => void
}
export type ProTableFieldProps<ValueType extends ProFieldValueType = ProFieldValueType> = Omit<
  ProFieldPropsType<ValueType>,
  'mode' | 'textProps' | 'modelValue' | 'renderFormItem' | 'render'
>
export type ProTableGetRowKey<RecordType = any> = (record: RecordType) => KeyType
export type ProTableEditContext<T = AnyObject> = {
  formRef: Ref<ProFormInstance | undefined>
  editable: ComputedRef<ProTableEditProps | undefined>
  editingCell: Ref<KeyType | undefined, KeyType | undefined>
  editingRowKeys: Ref<KeyType[], KeyType[]>
  getRowKey: ComputedRef<ProTableGetRowKey<T>>
  childrenColumnName: string
  getData: () => T[]
  setData: (data: T[]) => void
  removeRecords: Ref<T[]>
  editRecords: Ref<T[]>
  internalIndex: ProTableProviderProps['internalIndex']
  rowSelectionProps: ComputedRef<ProTableRowSelectionAfter<T>>
  waterfall: boolean
  pageInfo: ProTablePageParams
  getFormInstanceByKey: (key: string, rowData: T) => any
  onCreate: (e: MouseEvent) => void
}

export type ProTableOperateMethods<T = any> = {
  edit: (rowData: T) => void
  delete: (rowData: T) => void
  cancel: (rowData: T) => void
  save: (rowData: T) => void
  add: (rowData?: T) => void
}

export type ProTableProviderProps = {
  emit: ProTableEmits
  pageInfo: ProTablePageParams
  internalIndex: {
    value: number
  }
  originColumns: ComputedRef<ProColumns>
  columns: Ref<ProColumns>
  setTableColumns: (columns: ProColumns) => void
  resetTableColumns: () => void
  actions: {
    add: ProTableOperateMethods['add']
    delete: ProTableOperateMethods['delete']
  }
  setEditRecord: (v: any) => void
  waterfall: boolean
  getRowKey: ComputedRef<ProTableGetRowKey<any>>
  pureFetch: (p: ProTablePageParams) => Promise<any>
  setCellRef: (key: KeyType, ref: any) => void
  getCellRef: (key?: KeyType) => any
  getSelectionRows: () => AnyObject[]
  editableConfig: ComputedRef<ProTableEditProps | undefined>
  selectionRowKeys: Ref<KeyType[]>
  tableData: Ref<AnyObject[]>
  rowSelectionProps: ComputedRef<ProTableRowSelectionAfter<AnyObject>>
}

export type ProCellRendererParams = {
  cellKey: KeyType
  cellData: any
  rowData: any
  rowIndex: number
  columns?: ProColumns
  column: ProColumn
  columnIndex: number
}
export type KeyType = PropertyKey
export type AnyObject = Record<PropertyKey, any>
export type ProTableButtonProps = Partial<ButtonProps> & {
  onClick?: (e: MouseEvent, rowData: any, asyncEvent: ProAsyncEventType) => void
}
export type ProActionButtonProps =
  | false
  | ProTableButtonProps
  | ((params: ProCellRendererParams) => false | ProTableButtonProps)
// ...
export type ProTablePageParams = {
  currentPage?: number
  pageSize?: number
  total?: number
  scrollId?: string
  hasMore?: boolean
}
// ...
export type ProCellRenderer = (params: any) => VNode
export type ProHeaderCellRenderer = (params: any) => VNode
export type ProColumn = Omit<TextSpecifiledProps, 'copyText'> & {
  /** 标题，同label */
  title?: string
  /** 标题，同 title */
  label?: string
  /** 字段名 */
  field?: string
  /** 字段名，同 field */
  prop?: string
  /** 数据字段名，同 field */
  dataKey?: string
  /** 类型，支持 seq(序号) */
  type?: string
  /** 对齐方式 */
  align?: 'center' | 'right'
  /** column 模式 */
  mode?: string
  /** 控件类型 */
  valueType?: string
  /** 控件 props */
  proFieldProps?: AnyObject
  /** 是否必填 */
  required?: boolean
  /** 控件 校验规则 */
  rules?: PropFormFieldProps['rules'] | ((rowIndex: string) => PropFormFieldProps['rules'])
  /** 控件联动规则 */
  effects?: (rowIndex: string) => PropFormFieldProps['effects']
  /** cell自定义渲染 */
  render?: (scope: AnyObject, index: number) => VNode
  /** 在 搜索 中隐藏 */
  hideInSearch?: boolean
  /**在 弹窗表单 中删除 */
  hideInForm?: boolean
  /** 在 table 中隐藏 */
  hideInTable?: boolean
  /**
   *  @name 可编辑表格是否可编辑
   *
   * @example 不允许编辑
   * editable={false}
   *
   * @example 如果id=1不允许编辑
   * editable={(params)=> params.rowData.id !==1 }
   */
  editable?: false | ((params: ProCellRendererParams) => boolean)
  /** cell自定义渲染 */
  cellRenderer?: ProCellRenderer
  /** 编辑状态cell自定义渲染 */
  editCellRenderer?: ProCellRenderer
  /** 表头自定义渲染 */
  headerCellRenderer?: ProHeaderCellRenderer
  /** 嵌套表头 */
  children?: ProColumn[]
} & {
  [key: string]: any
}
export type ProColumns = ProColumn[]
// 编辑配置
type ProTableEditType = 'single' | 'multiple' | 'cell' | 'only'
export type ProTableEditProps = {
  type: ProTableEditType
  /** 是否隐藏操作列 */
  hideColumn?: boolean
  /** 新增时是否添加默认的一行 */
  defaultRow?: boolean
  /** 开启键盘 enter 跳转下一个 */
  keyboard?: boolean
  /** 定义 column 列回车切换下一行，若没有下一行则新增一行，默认最后一列 */
  keyboardNextRow?: string
  /** 保存一行的文字 */
  saveText?: string | VNode
  /** 取消编辑一行的文字 */
  cancelText?: string | VNode
  /** 删除一行的文字 */
  deleteText?: string | VNode
  /** 编辑一行的文字 */
  editText?: string | VNode
  /**
   * 编辑按钮属性
   */
  editButtonProps?: ProActionButtonProps
  /**
   * 删除按钮属性
   */
  deleteButtonProps?: ProActionButtonProps
  cellProps?: {
    mode?: 'doubleClick' | 'click'
  }
  deletePopconfirmMessage?: string
  onlyOneLineEditorAlertMessage?: string | VNode
  onlyAddOneLineAlertMessage?: string | VNode
  /**
   * 操作列属性
   */
  optionColumnProps?: any
}
/** 选中相关定义 */
export type ProTableRowSelectionProps<T> = {
  type?: 'checkbox' | 'radio'
  selectedRowKeys: KeyType | KeyType[]
  hideSelectAll?: boolean
  reserveSelection?: boolean
  rowClick?: boolean
  selectable?: (rowData: T) => boolean
  onChange?: (selectedRowKeys: KeyType[], selectedRows: T[]) => void
  onSelect?: (selectedRows: T[], row: T) => void | boolean | Promise<boolean>
}
export type ProTableRowSelectionData<T> = ProTableRowSelectionProps<T> & {
  selectedRowKeys: KeyType[]
}
// RowSelection 处理后的类型, selectable 一定存在
type ProTableRowSelectionAfter<T> = Omit<ProTableRowSelectionData<T>, 'selectable'> & {
  selectable: (row: T) => boolean
}
// ...
export type ProTableEmits<T = AnyObject> = {
  (e: 'menu-click', val: ProTableMenuClickParams<T>): void
}